Rename random_seed in Enrollment to commitment
We currently have a few concepts of "random seed"s.
There is `random_seed` in the Enrollment struct which is the
first preimage revealed. There is also `rand_seed` in
EnrollmentManager which is a preimage for a certain block height.

The first preimage has been renamed to commitment to reduce
ambiguity.